small update for Jordan

+++ src/main/java/com/qa/users/service/UserService.java
@@ -6,7 +6,7 @@ import com.qa.users.object.User;
 import com.qa.users.repository.UserRepo;
 
 @Service
-public class UserService implements UserInterface{
+public class UserService implements UserInterface {
 
 	private UserRepo repo;
 
@@ -16,25 +16,30 @@ public class UserService implements UserInterface{
 
 	public User readUser(long id) {
 		return repo.findById(id).get();
+		
+		
 	}
 
 	@Override
-	public String updatePassword(long id,String password) {
+	public String updatePassword(long id, String password) {
 		User aUser = readUser(id);
-		
-		if (aUser != null) {			
+
+		if (aUser != null) {
 			aUser.setPassword(password);
 			repo.save(aUser);
 			return aUser.getForename() + " " + aUser.getLastname() + "'s password has beeen changed.";
-		}return "No such user.";
-		
+		}
+		return "No such user.";
 
 	}
 
 	public String deleteUser(long id) {
-		String name = readUser(id).getForename()+" "+ readUser(id).getLastname();
-		this.repo.deleteById(id);
-		return name +"has been deleted.";
+		if (repo.findById(id) != null) {
+			String name = readUser(id).getForename() + " " + readUser(id).getLastname();
+			this.repo.deleteById(id);
+			return name + " has been deleted.";
+		}
+		return "No such user.";
 
 	}
 
@@ -45,8 +50,8 @@ public class UserService implements UserInterface{
 
 	public String updateRole(long id, String role) {
 		User aUser = readUser(id);
-		
-		if (aUser != null) {			
+
+		if (aUser != null) {
 			aUser.setRole(role);
 			repo.save(aUser);
 		}
